requirements.txt
Django==6.0.1
djangorestframework
Pillow
psycopg2-binary
gunicorn

fastapi
uvicorn[standard]
celery[redis]
redis
pytesseract
email-validator

pytest
pytest-cov
httpx
